requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bitsandbytes==0.42.0
boto3==1.34.20
botocore==1.34.20
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
cmake==3.25.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
diffusers==0.32.1
einops==0.8.0
environs==10.3.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.12.0
gradio==3.41.2
gradio_client==0.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
humanize==4.9.0
idna==3.4
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lit==15.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.10.0
moviepy==1.0.3
mpmath==1.3.0
multiprocessing-logging==0.3.4
narwhals==1.20.1
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
orjson==3.10.13
packaging==24.2
pandas==2.2.3
peft==0.14.0
petrel-oss-sdk===v2.2.1-2-g1505ef3-master
pillow==10.2.0
proglog==0.1.10
protobuf==3.20.0
psutil==6.1.1
py-cpuinfo==9.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytorch-wavelets==1.3.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.28.1
rpds-py==0.22.3
s3transfer==0.10.0
safetensors==0.4.5
scipy==1.15.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.67.1
transformers==4.47.1
triton==2.3.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==1.26.13
uvicorn==0.34.0
websockets==11.0.3
zipp==3.21.0
